# apps/dashboard-frontend/specs/openapi-control.json
{"openapi": "3.1.0", "info": {"title": "OpenRouter Control Plane", "version": "0.1.0"}, "paths": {"/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/auth/login": {"post": {"summary": "Login", "operationId": "login_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/register": {"post": {"summary": "Register", "operationId": "register_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api-keys/create": {"post": {"summary": "Create New Key", "operationId": "create_new_key_api_keys_create_post", "parameters": [{"name": "name", "in": "query", "required": true, "schema": {"type": "string", "title": "Name"}}, {"name": "user_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api-keys": {"get": {"summary": "List Api Keys", "operationId": "list_api_keys_api_keys_get", "parameters": [{"name": "user_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/models": {"get": {"summary": "List Models", "operationId": "list_models_models_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ModelResponse"}, "type": "array", "title": "Response List Models Models Get"}}}}}}}, "/users/{user_id}": {"get": {"summary": "Get User Status", "operationId": "get_user_status_users__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/{user_id}/provider-keys": {"post": {"summary": "Add Provider Key", "operationId": "add_provider_key_users__user_id__provider_keys_post", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderKeyRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"summary": "List Provider Keys", "operationId": "list_provider_keys_users__user_id__provider_keys_get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/{user_id}/provider-keys/{provider_name}": {"delete": {"summary": "Delete Provider Key", "operationId": "delete_provider_key_users__user_id__provider_keys__provider_name__delete", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}, {"name": "provider_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Provider Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/analytics/usage": {"get": {"summary": "Get Usage Stats", "description": "Returns real token usage over the last 7 days for the chart.", "operationId": "get_usage_stats_analytics_usage_get", "parameters": [{"name": "user_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/analytics/health": {"get": {"summary": "Get Provider Health", "description": "The 'Weather Map' API.", "operationId": "get_provider_health_analytics_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/analytics/preference": {"post": {"summary": "Save User Preference", "description": "Saves the RLHF data for future model training.", "operationId": "save_user_preference_analytics_preference_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PreferenceRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"AuthRequest": {"properties": {"email": {"type": "string", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "AuthRequest"}, "CompanyResponse": {"properties": {"name": {"type": "string", "title": "Name"}, "website": {"type": "string", "title": "Website"}}, "type": "object", "required": ["name", "website"], "title": "CompanyResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MappingResponse": {"properties": {"provider": {"type": "string", "title": "Provider"}, "input_token_cost": {"type": "number", "title": "Input Token Cost"}, "output_token_cost": {"type": "number", "title": "Output Token Cost"}}, "type": "object", "required": ["provider", "input_token_cost", "output_token_cost"], "title": "MappingResponse"}, "ModelResponse": {"properties": {"id": {"type": "integer", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "slug": {"type": "string", "title": "Slug"}, "context_length": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Context Length"}, "company": {"anyOf": [{"$ref": "#/components/schemas/CompanyResponse"}, {"type": "null"}]}, "mappings": {"items": {"$ref": "#/components/schemas/MappingResponse"}, "type": "array", "title": "Mappings", "default": []}}, "type": "object", "required": ["id", "name", "slug"], "title": "ModelResponse"}, "PreferenceRequest": {"properties": {"prompt": {"type": "string", "title": "Prompt"}, "primary_model": {"type": "string", "title": "Primary Model"}, "primary_response": {"type": "string", "title": "Primary Response"}, "shadow_model": {"type": "string", "title": "Shadow Model"}, "shadow_response": {"type": "string", "title": "Shadow Response"}, "user_preference": {"type": "string", "title": "User Preference"}}, "type": "object", "required": ["prompt", "primary_model", "primary_response", "shadow_model", "shadow_response", "user_preference"], "title": "PreferenceRequest"}, "ProviderKeyRequest": {"properties": {"provider_name": {"type": "string", "title": "Provider Name"}, "api_key": {"type": "string", "title": "Api Key"}}, "type": "object", "required": ["provider_name", "api_key"], "title": "ProviderKeyRequest"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}