# apps/dashboard-frontend/src/client-control/models/PreferenceRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PreferenceRequest = {
    prompt: string;
    primary_model: string;
    primary_response: string;
    shadow_model: string;
    shadow_response: string;
    user_preference: string;
};